requirements.txt
biopython==1.79
click==8.1.3
contextlib2==21.6.0
colorlog==6.7.0
dm-haiku==0.0.4
jax==0.2.14
jaxlib==0.1.73+cuda11.cudnn82
ml_collections==0.1.1
mock==5.1.0
numpy==1.23.5
protobuf==3.19.0
scipy==1.9.3
tensorflow==2.10.0
torch==1.13.0
yacs==0.1.8